Move parser interfaces & implementation to own file
